AutoUpdateInfo: fix equals method to use String equals
This change has not been tested, but the implementation of
the equals method for AutoUpdateInfo seems wrong, also when
one look at the hashCode method.

So instead of using == for comparing strings, use the String.equals
method.

I am guessing that the current implementation of the equals may be
causing duplicate files to be uploaded, ref problems in pull request
number 43, but this is highly speculative from my side.

I've changed the implementation of the equals method to be like
the other overriden equals methods in the project.

If my changes are wrong, the equals method in AutoUpdateInfo should
receive some comments as to why the use of == is correct