SeafileObserver: Remove files from watched and recent lists
After a file is no longer recent, it should be removed from the recent downloaded list,
to avoid that list growing forever larger.
When a file has been removed from the device, it should no longer be in the list of
watched files, to avoid that list growing forever larger.