Enable TLSv1.1+, disable SSLv3
This patch forces Seadroid to use a specific list of TLS protocols.
Specifically, SSLv3 is disabled, as it is considered insecure by now.
Also TLSv1.1+ are enabled (they are available with API 16+ but disabled
up to API 19).

To do so, we have to inherit from SSLSocketFactory, which makes the codes
somewhat verbose. The upside is that this class can later be used to further
modify the Android defaults concerning SSL (e.g. cyphersuites used).

This patch also fixes a NPE when connecting to a server that only supports
TLSv1.2.

References:
* https://developer.android.com/reference/javax/net/ssl/SSLSocket.html#getSupportedCipherSuites%28%29