Specifiy TLS cipher list
By default Android will offer the HTTPS server ciphers which are today
considered to be insecure (e.g. SSL_DHE_DSS_EXPORT_WITH_DES40_CBC_SHA).
A properly configured server should reject those ciphers, but better be
on the safe side on not offer those ciphers by the client at all.

Additionally the order of the cipher matters as well, as more secure/fast
ciphers should be at the top of the list.

This patch adds a cipher list based on
https://briansmith.org/browser-ciphersuites-01.html

The list has to be revised from time to time as the security of TLS
is re-evaluated by security research.

References:
* https://briansmith.org/browser-ciphersuites-01.html
* https://developer.android.com/reference/javax/net/ssl/SSLSocket.html