fix: error when uploading same file again
Reasoning: Temp files aren't deleted after upload.
If the user tries to upload the same file twice,
the temp file cannot be created again, causing an

    throw new RuntimeException("could not create temporary file");

We can't delete or overwrite the existing temp file, as the upload
might still be in progress.

There are other ways to fix this issue, but this is the simplest solution,
as far as I can see.