instruct android gallery to ignore files locally cached by seafile
this affects

- the Android Gallery
- the ACTION_GET_CONTENT image picker

Both will crawl the seafile internal directories from time to time
and include the containing images into the gallery. I think this is
confusing. The gallery will only show the small subset of cached images,
which is not very useful and potentially confusing to users.

Also the way Seafile currently purges its cache, the Android gallery won't
get notified if images are deleted(*). This results in stale entries in the
image gallery of files. The files are listed (sometimes with thumbnails,
sometimes not). But trying to open them will result in an error.

(*) This could be fixed, however, in other ways.