SeafileProvider: fix: hang during thumbnail download
sometimes the SAF file chooser can choke up when displaying high resolution
thumbnails (like on a tabled in grid view). The reason was that we tried
to write into a pipe with an unconected end. the pipe buffer would run fill
and then block.

Fix this by returning the file descriptor early, and then agressively close
the pipe if we get a cancellation signal.