SeafileProvider: disable access if server is unreachable
If the server is unreachable, we don't allow download of any files.
To make this fact more clear to the user, grey out all entries in an
Seafile account.

The reachability test is done for every server individually.

This should mostly avoid issues of users trying to download files from unavailable
servers.