SeafileProvier: fix: refresh directory
change the way we avoid infinite refresh loops.
previously the user had to change the directory to trigger
a refresh. this is problematic if there is only one directory
(the list of repos). the user will never be able to trigger a refresh.

this fixes the following issue: There is only one server, and that
server is greyed out due to unreachability. Then the user won't be
able to get SeafileProvider to retest if the server has become available
again.