fix crash when open .kdb and .kdbx files
1. All MIME types received from outside Android
(such as user input, or external sources like Bluetooth, NFC, or the Internet)
should be normalized before they are used to create an Intent.
2. MIME type and Uri scheme matching in the Android framework is case-sensitive, unlike the formal RFC definitions.
As a result, you should always write these elements with lower case letters,
or use normalizeMimeType(String) or normalizeScheme() or setDataAndTypeAndNormalize(Uri, String)
to ensure that they are converted to lower case.
3. use setDataAndTypeAndNormalize instead of setDataAndType