redesign and rewrite notification
1. All downloading events will be represented by one downloading notification, the same with uploading notification.
2. To update notification, simply call its notificationID.
The update events will be invoked by UploadTaskManager and DownloadTaskManager
because they implemented the UploadStateListener and DownloadStateListener respectively.
3. It always failed to calculate total size by transferService,
because file size of each DownloadTask instance needs to be fetched on first progress update,
that will fail when multi tasks waiting in a queue because first progress update won\`t get called instantly.
As a workaround, the total size was calculated by seafDirent.
4. If the notification provider exist, update its data with the same notificationID,
remember to update totalSize and taskIDList.
If the notification provider doesn\`t exist, create a new one and cache it for later use.
5. Supplying the onging notification to be shown to the user while transferring files
in order to keep running at the foreground.
By default services are background, meaning that if the system needs to kill them to reclaim more memory, they can be killed without too much harm.