calculate progress by transferService
It is not the best way to calculate total size by transferService,
because we can not get an accurate value in this way.
We know file size only when the relavant download task under executing,
that will fail when multi tasks waiting in a queue.

An alternative way are expected in the future.