keep photos being uploaded continuously
The system keeps all network connections open while in background, but does not deliver the data until the app resumes.
In order to ensure continuous network communication for camera uploading,
put the service in a foreground state, where the system considers it to be something the user is actively aware of
and thus not a candidate for killing when low on memory.

TransferService was set to be running in the foreground instead of CameraUploadService.
CameraUploadNotificationProvider was used to manage & update notification