Revert "use native way to load drawables"
This patch fixes the following bug:

When scrolling images rapidly, the wrong thumbnails will be displayed for up to
a second.

steps to reproduce:
1. enter a directory with images
2. start scrolling before all thumbnails are loaded.
3. stop scrolling.
4. you will see the wrong thumbnails be displayed for a short while.

The reason for this is that ImageLoader makes the assumption that *all* access to
the ImageView is done through the ImageLoader. Bypassing the ImageLoader and
setting a Drawable directly will mess up the internal state keeping of the ImageLoader.

This reverts commit 60bdb83c54397bdcd8b1c8e896f6e82318c3e331.