bugfix: early SAF usage by app might fail
Previously it was assumed that SeafileProvider.queryRoots() would be called
before any of the other SeafileProvider.queryXXX() methods. That was a mistake.
At least SeafileProvider.queryDocument() might actually be called earlier.

This can cause problems if the Android OS has just been freshly rebooted and
the SAF is used for the first time. In that case

    SeafileProvider.isReachable

will still be empty. Thus doing an

    SeafileProvider.isReachable.get(account)

will return null. That breaks thinks, as can be witnessed e.g. by trying to use
latest Keepass2Android release with Seadroid after a fresh reboot.

This patches fixes that issue by converting the HashMap into a Set
(which is a more fitting data structure anyway).

(Making Account "Comparable" is just necessary to make the ConcurrentSkipListSet
work.)