Remember browse position when re-entering Seadroid via Launcher icon
When tapping on the Seadroid Launcher icon, Seadroid currently always reverts back to
its "Home screen" (the list of repositories). Even if Seadroid was
already running and the user had already entered a specific folder.

This behaviour differs from most other apps. Usually the state of the app is preserved
if one re-enters it via the Launcher app.

This patch tries to do exactly that. The key place is BrowserActivity.onNewIntent()
In this method we have to decide whether to kill the old BrowserActivity and create a new
one; or to just keep the old one running.

Previously this was decided on the given account information. However this does not work
with the Launcher, as the account info is null if the intent comes from the Launcher.

I decided to use the Intent category/action for making that decision. If it is a MAIN
action, than it was called by the Launcher and the old instance shall be retained.

Therefore I had to convert UnlockGesturePasswordActivity and AccountsActivity to call
the BrowseActivity directly by its class name and not the action.

Note that a shorter fix would have been:

    -        if (!account.equals(selectedAccount)) {
    +        if (server == null || !account.equals(selectedAccount)) {

However BrowseActivity is started from a lot places, and I was a bit scared of creating new
bugs. This patch should hopefully have more controlled side effects.