Open file in new task, keep BrowseActivity alive
Currently the BrowseActivity finishes itself after opening a file.
So if a user had to do a lot of business is a specific directory, she
would have to enter that directory again and again.

Better to start the ACTION_NEW in a new task and keep Seadroid around
in the background.