AccountDetailActivity should not start BrowserActivity
this leads to unwanted activity stacks.

Scenario 1: (switching to another account)

BrowserActivity
 -> AccountsActivity
    -> AccountDetailActivity

when the user then logins in, the new stack will be:

BrowserActivity
 -> AccountsActivity
    -> BrowserActivity

... and that's not good.

Scenario 2: (entering an account for the first time)

AccountsActivity
  -> AccountDetailActivity

when the user then logins in, the new stack will be:

AccountsActivity
  -> BrowserActivity

.. again, not good.

So, better to return from the AccountDetailActivity back to the AccountActivity,
and start the BrowserActivity from there.