Seadroid is not ready yet for API level 23
API 23 introduces runtime permissions:

https://developer.android.com/about/versions/marshmallow/android-6.0-changes.html#behavior-runtime-permissions

Once an app targets API 23, the user isn't shown the list of required permissions
during install. Instead, the app has to check and ask the user for
permissions on first use (checkSelfPermissions() / requestPermissions()).

At the moment, Seadroid is not implementing this new permissions API. So
targeting Seadroid for API level 23 would break it on Android 6.0 devices
when newly installed.