SeafileProvider: assume at beginning all accounts are reachable
This is to avoid problems right after boot. If an app like Titanium Backup
used the SeafileProvider for the first time after a reboot, then we don't know
yet if the server is reachable. However we should assume that it is reachable.
Otherwise we would return an error to the app and confuse it.