[PATCH 1/3] storage: Paths saved in REPODIR_TABLE_NAME must be relative
Currently, absolute paths are saved for each repository in the REPODIR_TABLE.

This didn't matter in the past, aslong as Seadroid had a fixed cache directory.
However, the next patch in this patchset will allow users to move the directory
to other storage devices.

This patch increments the Database Version, since the meaning of
REPODIR_COLUMN_REPO_DIR is changed. While at it, also remove the column
REPODIR_COLUMN_REPO_NAME, as it is unneccessary (REPO_ID is a sufficient key).