[PATCH 3/3] storage: Use private cache for JSON files
Currently, JSON files containing listings of repositories and files/directories
are stored in external memory, where many other apps have read/write access to.
This is a potential privacy issue. The user will likely understand that
downloaded/cached files will be accessible by other apps. However, the fact
that directory listings and the list of repositories on a server are silently
accessible by other apps as well, that is not obvious and should therefore not
be allowed.

This patch changes the storage path for Seadroid JSON files to the internal
cache directory. Only Seadroid (and root-apps) have access to that directory.