Camera Upload: improve check for existing files
Camera Upload tries to avoid creating duplicates when uploading files.
Previously, only the exact file name was checked for existence on the server.
This failed in situations where the file was already existing under a
duplicate name, like "DSC0001 (1).jpg".

This patch addresses that by using a regex to compare the filenames.

This fixes issue #503