Camera upload: enabling global sync on android
this enables global sync on android whenever seadroid camera upload is enabled.

This is a quick hack and should never be merged! Its just to check if this might
work on certain custom ROM devices.

If this actually enables global sync, we have to add some crucial UI. This guy makes
some good comments regarding that: http://stackoverflow.com/a/11524302