Implement biometric lock
This adds a simple biometric lock to the whole application. Once toggled
in settings, it requires a fingerprint/password/other authentication
method to be used before the app can be opened.

Until it's closed, the app will stay permanently unlocked. This doesn't
implement locking when moving to a different window / after a certain
period of time, account-specific locks, or wiping auth data when
biometrics change.